requirements.txt
Flask>=2.2
requests>=2.28
python-dotenv>=1.0